One more place to prevent overflow

+++ src/bmp.imageio/bmpinput.cpp
@@ -396,7 +396,8 @@ BmpInput::read_native_scanline(int subimage, int miplevel, int y, int /*z*/,
          || m_dib_header.compression == RLE8_COMPRESSION)
         && m_colortable.size()) {
         for (int x = 0; x < m_spec.width; ++x) {
-            int p = m_uncompressed[(m_spec.height - 1 - y) * m_spec.width + x];
+            int p = m_uncompressed[int64_t(m_spec.height - 1 - y) * m_spec.width
+                                   + x];
             auto& c              = colortable(p);
             mscanline[3 * x]     = c.r;
             mscanline[3 * x + 1] = c.g;
